Windows Update pullpatches.bat April edition
As I'm the only user, I've removed all patches except for
W7 64 bit. Windows 8 is elided. Besides, W8 became more
complex with extra patches just for it...

IE11 is a cumulative update. Not sure if that means it actually
supersedes the March IE11 patch so I've left both patchsets in
for now.